requirements.txt
streamlit
opencv-python
numpy
Pillow
scikit-learn
joblib
scikit-image